requirements.txt
scholarly >= 1.1.0
pyyaml >= 5.4.1